+++ src/pages/SignIn/SignIn.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { styled } from "styled-components";
-import Signinform from "~components/auth/Signin/Signinform";
+import Signinform from "~pages/SignIn/Signinform";
 
 function SignIn() {
   return <Signinform />;
