+++ src/utils/constant.js
@@ -3,4 +3,5 @@ export const ROUTER_LINKS = {
   signin: "/signin",
   signup: "/signup",
   items: "/items",
+  additem: "/additem",
 };
